+++ tests/unit/test_bruteforce.cpp
@@ -1,6 +1,5 @@
 #include "gtest/gtest.h"
 #include "VecSim/vec_sim.h"
-#include "VecSim/utils/vec_utils.h"
 #include "test_utils.h"
 
 class BruteForceTest : public ::testing::Test {
@@ -335,44 +334,6 @@ TEST_F(BruteForceTest, test_bf_info) {
     ASSERT_EQ(info.bfInfo.blockSize, 1);
     ASSERT_EQ(info.bfInfo.indexSize, 0);
     VecSimIndex_Free(index);
-}
-
-void compareFlatIndexInfoToIterator(VecSimIndexInfo info, VecSimInfoIterator *infoIter) {
-    ASSERT_EQ(7, VecSimInfoIterator_NumberOfFields(infoIter));
-    while (VecSimInfoIterator_HasNextField(infoIter)) {
-        VecSim_InfoField *infoFiled = VecSimInfoIterator_NextField(infoIter);
-        if (!strcmp(infoFiled->fieldName, VecSimCommonStrings::ALGORITHM_STRING)) {
-            // Algorithm type.
-            ASSERT_EQ(infoFiled->fieldType, INFOFIELD_STRING);
-            ASSERT_STREQ(infoFiled->stringValue, VecSimAlgo_ToString(info.algo));
-        } else if (!strcmp(infoFiled->fieldName, VecSimCommonStrings::TYPE_STRING)) {
-            // Vector type.
-            ASSERT_EQ(infoFiled->fieldType, INFOFIELD_STRING);
-            ASSERT_STREQ(infoFiled->stringValue, VecSimType_ToString(info.bfInfo.type));
-        } else if (!strcmp(infoFiled->fieldName, VecSimCommonStrings::DIMENSION_STRING)) {
-            // Vector dimension.
-            ASSERT_EQ(infoFiled->fieldType, INFOFIELD_UINT64);
-            ASSERT_EQ(infoFiled->uintegerValue, info.bfInfo.dim);
-        } else if (!strcmp(infoFiled->fieldName, VecSimCommonStrings::METRIC_STRING)) {
-            // Metric.
-            ASSERT_EQ(infoFiled->fieldType, INFOFIELD_STRING);
-            ASSERT_STREQ(infoFiled->stringValue, VecSimMetric_ToString(info.bfInfo.metric));
-        } else if (!strcmp(infoFiled->fieldName, VecSimCommonStrings::INDEX_SIZE_STRING)) {
-            // Index size.
-            ASSERT_EQ(infoFiled->fieldType, INFOFIELD_UINT64);
-            ASSERT_EQ(infoFiled->uintegerValue, info.bfInfo.indexSize);
-        } else if (!strcmp(infoFiled->fieldName, VecSimCommonStrings::BLOCK_SIZE_STRING)) {
-            // Block size.
-            ASSERT_EQ(infoFiled->fieldType, INFOFIELD_UINT64);
-            ASSERT_EQ(infoFiled->uintegerValue, info.bfInfo.blockSize);
-        } else if (!strcmp(infoFiled->fieldName, VecSimCommonStrings::MEMORY_STRING)) {
-            // Memory.
-            ASSERT_EQ(infoFiled->fieldType, INFOFIELD_UINT64);
-            ASSERT_EQ(infoFiled->uintegerValue, info.bfInfo.memory);
-        } else {
-            ASSERT_TRUE(false);
-        }
-    }
 }
 
 TEST_F(BruteForceTest, test_basic_bf_info_iterator) {
